requirements.txt
opencv-python
matplotlib
pillow
numpy
opencv-python-headless
ultralytics